fix-dto: wallet dtos completed
