--- backend/src/app/dtos/wallet/add-money.dto.ts
@@ -1,4 +1,0 @@
-export interface AddMoneyDto {
-  amount: number;
-  currency?: string;
-}
