--- backend/src/app/dtos/wallet/reduce-money.dto.ts
@@ -1,4 +1,0 @@
-export interface ReduceMoneyDto {
-  amount: number;
-  currency?: string;
-}
