--- backend/src/app/dtos/wallet/wallet-response.dto.ts
@@ -1,8 +1,0 @@
-export interface WalletResponseDto {
-  id: string;
-  userId: string;
-  balance: number;
-  currency: string;
-  createdAt: Date;
-  updatedAt: Date;
-}
