+++ backend/src/app/usecases/wallet/implementation/get-wallet.usecase.ts
@@ -1,7 +1,8 @@
-import { WalletResponseDto } from "../../../dtos/wallet/wallet-response.dto";
+
 import { IWalletRepository } from "../../../repositories/wallet.repository.interface";
 import { IGetWalletUseCase } from "../interfaces/get-wallet.usecase.interface";
 import { Wallet } from "../../../../domain/entities/wallet.entity";
+import { WalletResponseDto } from "@/app/dtos/wallet";
 
 export class GetWalletUseCase implements IGetWalletUseCase {
   constructor(private readonly walletRepository: IWalletRepository) {}
