+++ backend/src/presentation/validators/wallet.ts
@@ -1,8 +1,6 @@
-import { z } from "zod";
+import z from "zod";
 
 export const TopUpWalletDtoSchema = z.object({
   amount: z.number().positive(),
   paymentMethod: z.enum(["WALLET", "STRIPE", "PAYPAL", "RAZORPAY"]),
 });
-
-export type TopUpWalletDto = z.infer<typeof TopUpWalletDtoSchema>;
